+++ grappelli/static/grappelli/js/actions.js
@@ -1,5 +1,0 @@
-/**
- * GRAPPELLI ACTIONS.JS
- * see actions.min.js
- * 
- */