+++ docs/changelog.rst
@@ -8,6 +8,22 @@
 Changelog
 =========
 
+2.4.0 (July 20, 2012)
+---------------------
+
+* Compatibility with Django 1.4
+* New: Grappelli namespace (grp-) for css-classes.
+* New: Grappelli namespace (grp.jQuery) for jQuery.
+* New: Compass-based CSS
+* New: Added toggle-all with change-form
+* New: DOM-Documentation
+* New: ``sortable_excludes``
+* New: 2 different Changelists (one with sidebar filters).
+* New: Minified Grappelli Javascripts.
+* New: Added support for django-reversion.
+* Changed: Selectors with Inlines in order to update the sortable-field (now also works with File-Fields)
+* Changed: Updated jQuery to 1.7.2
+
 2.3.8 (April 03, 2012)
 ----------------------